@@ -1,8 +1,9 @@
 package com.fbsum.android.adapterdelegates.sample;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.fbsum.android.adapterdelegates.sample.model.ComplexItem;
 import com.fbsum.android.adapterdelegates.sample.model.ContentItem;
