@@ -1,10 +1,11 @@
 package com.fbsum.android.adapterdelegates.sample;
 
 import android.app.Activity;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.fbsum.android.adapterdelegates.SimpleAdapterDelegate;
 import com.fbsum.android.adapterdelegates.sample.model.ContentItem;
