@@ -16,8 +16,9 @@
 
 package com.fbsum.android.adapterdelegates;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.ViewGroup;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.List;
 
